Common DeletePolicy
